вынести валидацию и лимитер в отдельные модули
