+++ middlewares/validations.js
@@ -1,0 +1,46 @@
+const { celebrate, Joi } = require('celebrate');
+const { reg } = require('../utils/constants');
+
+module.exports.createUserValidation = celebrate({
+  body: Joi.object().keys({
+    name: Joi.string().min(2).max(30),
+    email: Joi.string().required().email({ minDomainSegments: 2 }),
+    password: Joi.string().required(),
+  }),
+});
+
+module.exports.loginUserValidation = celebrate({
+  body: Joi.object().keys({
+    email: Joi.string().required().email({ minDomainSegments: 2 }),
+    password: Joi.string().required(),
+  }),
+});
+
+module.exports.updateMyProfileValidation = celebrate({
+  body: Joi.object().keys({
+    name: Joi.string().min(2).max(30).required(),
+    email: Joi.string().email({ minDomainSegments: 2 }).required(),
+  }),
+});
+
+module.exports.createMovieValidation = celebrate({
+  body: Joi.object().keys({
+    country: Joi.string().required(),
+    director: Joi.string().required(),
+    duration: Joi.number().required(),
+    year: Joi.string().required(),
+    description: Joi.string().required(),
+    image: Joi.string().required().regex(reg),
+    trailerLink: Joi.string().required().regex(reg),
+    thumbnail: Joi.string().required().regex(reg),
+    movieId: Joi.number().required(),
+    nameRU: Joi.string().required(),
+    nameEN: Joi.string().required(),
+  }),
+});
+
+module.exports.deleteMovieValidation = celebrate({
+  params: Joi.object().keys({
+    movieId: Joi.string().required().length(24).hex(),
+  }),
+});
