+++ routes/users.js
@@ -1,18 +1,13 @@
 const userRouter = require('express').Router();
-const { celebrate, Joi } = require('celebrate');
 
 const {
   getMyProfile,
   updateMyProfile,
 } = require('../controllers/users');
+const updateMyProfileValidation = require('../middlewares/validations');
 
 userRouter.get('/me', getMyProfile);
 
-userRouter.patch('/me', celebrate({
-  body: Joi.object().keys({
-    name: Joi.string().min(2).max(30).required(),
-    email: Joi.string().email({ minDomainSegments: 2 }).required(),
-  }),
-}), updateMyProfile);
+userRouter.patch('/me', updateMyProfileValidation, updateMyProfile);
 
 module.exports = userRouter;
